NavigationLeftTltle作成のextensionコードと実装
